# src/mcp/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task.sh MCP Server",
    "description": "Model Context Protocol server for Task.sh - Provides task context and export tools for LLMs",
    "version": "1.0.0",
    "contact": {
      "name": "Task.sh API Support"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Development MCP Server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_KEY",
        "description": "API key for authenticating with the MCP server"
      }
    },
    "schemas": {
...
      "ToolResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "tool_result": {
            "type": "object",
            "properties": {
              "format": {"type": "string"},
              "content_type": {"type": "string"},
              "file_extension": {"type": "string"},
              "data": {},
              "filename": {"type": "string"}
            }
          },
          "metadata": {"type": "object"}
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean", "enum": [false]},
          "error": {"type": "string"},
          "code": {"type": "string"},
          "message": {"type": "string"}
        }
      }
    }
  },
  "paths": {
...
    "/tools/export/project": {
      "post": {
        "summary": "Export specific project",
        "description": "Exports tasks from a specific project with optional filtering",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["project_name"],
                "properties": {
                  "project_name": {"type": "string"},
                  "format": {"type": "string", "enum": ["json", "csv", "markdown", "excel"], "default": "json"},
                  "status": {"type": "string", "enum": ["complete", "incomplete", "all"], "default": "all"},
                  "date_from": {"type": "string", "format": "date"},
                  "date_to": {"type": "string", "format": "date"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project exported successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ToolResponse"}
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
    "/tools/export/status": {
      "post": {
        "summary": "Export tasks by status",
        "description": "Exports all tasks filtered by completion status",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {"type": "string", "enum": ["complete", "incomplete"]},
                  "format": {"type": "string", "enum": ["json", "csv", "markdown", "excel"], "default": "json"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status export completed successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ToolResponse"}
              }
            }
          }
        }
      }
    },
    "/tools/export/date-range": {
      "post": {
        "summary": "Export tasks by date range",
        "description": "Exports tasks within a specified date range with optional filtering",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["date_from", "date_to"],
                "properties": {
                  "date_from": {"type": "string", "format": "date"},
                  "date_to": {"type": "string", "format": "date"},
                  "format": {"type": "string", "enum": ["json", "csv", "markdown", "excel"], "default": "json"},
                  "project": {"type": "string"},
                  "status": {"type": "string", "enum": ["complete", "incomplete", "all"], "default": "all"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Date range export completed successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ToolResponse"}
              }
            }
          }
        }
      }
    },
    "/tools/export/bulk": {
      "post": {
        "summary": "Bulk export operations",
        "description": "Performs multiple export operations in a single request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "projects": {"type": "array", "items": {"type": "string"}},
                  "status_filters": {"type": "array", "items": {"type": "string", "enum": ["complete", "incomplete"]}},
                  "date_ranges": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "start": {"type": "string", "format": "date"},
                        "end": {"type": "string", "format": "date"},
                        "label": {"type": "string"}
                      }
                    }
                  },
                  "format": {"type": "string", "enum": ["json", "csv", "markdown", "excel"], "default": "json"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk export completed successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ToolResponse"}
              }
            }
          }
        }
      }
    }
  },
...
}